requirements.txt
# --- NÚCLEO API ---
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
python-multipart>=0.0.9

# --- UTILIDADES ---
requests>=2.31.0
httpx>=0.26.0
typing-extensions>=4.9.0

# --- BASES DE DATOS (Fase 2 Preparación) ---
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
# qdrant-client  <-- Lo activaremos cuando llegue el M.2
